--- prince-archiver/prince_archiver/data_archive/dto.py
@@ -1,15 +1,0 @@
-from datetime import date
-from uuid import UUID
-
-from pydantic import BaseModel
-
-
-class Archive(BaseModel):
-    path: str
-    src_keys: list[str]
-
-
-class UpdateArchiveEntries(BaseModel):
-    job_id: UUID
-    date: date
-    archives: list[Archive]
