+++ prince-archiver/prince_archiver/service_layer/external_dto.py
@@ -39,3 +39,15 @@ class TimestepDTO(TimestepMeta):
             root = ts.strftime("%Y%m%d_%H%M.tar")
             self.key = f"{self.experiment_id}/{root}"
         return self
+
+
+# Messages emitted from Surf Data Archive
+class Archive(BaseModel):
+    path: str
+    src_keys: list[str]
+
+
+class UpdateArchiveEntries(BaseModel):
+    job_id: UUID
+    date: date
+    archives: list[Archive]
